add vdb debugger backend.
Add a debugger backend (like GDB or LLDB support) for vdb (https://github.com/vivisect/vivisect). The major work was really only figuring out how to get the dbgentry working. It took a few hacks, but seems to work fairly stably. Had to add a custom lexer because 1) regexes are write-once, never read again, and I was having a bit of trouble reversing the LLDBIntelLexer, and 2) the vdb Intel listing has a column for raw hex bytes.